src/ToG-2/requirements.txt
accelerate==0.31.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.7.2
attrs==24.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.2
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blingfire==0.1.8
blis==0.7.11
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
datasets==3.0.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
et-xmlfile==1.1.0
fastapi==0.115.0
filelock==3.16.1
FlagEmbedding==1.2.11
flatbuffers==23.5.26
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.10.0
google-auth==2.25.2
google-search-results==2.4.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib-resources==6.1.1
importlib_metadata==8.5.0
interegular==0.3.3
isodate==0.6.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.3.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langcodes==3.4.0
langsmith==0.0.71
language_data==1.2.0
lark==1.2.2
llama-index==0.9.40
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.3.0
marisa-trie==1.2.0
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.9.1.post1
mistral_common==1.4.4
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.50.2
openpyxl==3.1.5
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
outcome==1.3.0.post0
outlines==0.0.46
overrides==7.4.0
packaging==23.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
peft==0.12.0
pillow==10.4.0
playwright==1.46.0
posthog==3.1.0
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
protobuf==4.23.4
psutil==6.0.0
pulsar-client==3.3.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.1.0
pyparsing==3.1.1
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
ray==2.37.0
rdflib==7.0.0
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.26.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
serpapi==0.1.5
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.23
srsly==2.4.8
starlette==0.38.6
sympy==1.13.3
tenacity==8.2.3
thinc==8.2.5
threadpoolctl==3.2.0
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.45.1
trio==0.27.0
trio-websocket==0.11.1
triton==3.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.16
uvicorn==0.31.0
uvloop==0.20.0
vllm==0.6.2
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.27.post2
xxhash==3.5.0
yagooglesearch==1.10.0
yarl==1.13.1
zipp==3.20.2
